Create Predict Students' Dropout and Academic Success
This dataset was created to address the challenge of academic dropout and failure in higher education. It aims to identify students at risk early in their academic journey using machine learning techniques. The dataset contains information collected at the time of student enrollment, including academic path, demographics, and socio-economic factors. The problem is framed as a three-category classification task, classifying students as dropout, enrolled, or graduate by the end of the standard course duration.

Funding for this dataset comes from the SATDAP program - Capacitação da Administração Pública under grant POCI-05-5762-FSE-000191 in Portugal.

Each instance in the dataset represents a student. It is recommended to split the data into 80% for training and 20% for testing.

Rigorous data preprocessing has been performed to handle anomalies, outliers, and missing values. Notably, there are no missing values in the dataset.