Create Project Description
The purpose of classification machine learning in this context is to leverage the dataset's information to predict the outcome of students' academic journeys. By classifying students into categories such as dropout, enrolled, or graduate, machine learning algorithms can assist in identifying students who are at risk of academic failure or dropout at an early stage. This predictive capability enables educational institutions to intervene and provide targeted support and interventions to at-risk students, ultimately contributing to the reduction of academic dropout rates and improving overall student success in higher education.